LayoutTests/imported/w3c/web-platform-tests/html/rendering/the-details-element/details-blockification-expected.txt

foo
bar
foo
bar

FAIL Summary and content should have display:block computed value assert_equals: expected "block" but got "inline"
